Improve keypress detection
From now we use `Window.get_keys_down()` instead of
`Window.get_keys_pressed()`. The reason is that it looks like
`get_keys_pressed()` only returns the new keys that have been pressed
since the last update. If a key has been held down for a while, the
repeat interval at which the key press is fired again is not fast enough
for a smooth experience. However, `get_keys_down()` returns all keys
that are currently down, which means that keys that are held down get
returned everytime until they are let go.